+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
-- [**BC**] Add `KeyForRenewToken` type to explicitly mark a JWTKey purpose
+- Add `KeyForRenewToken` type to explicitly mark a JWTKey purpose
+- Add `CurrentApplication` type to explicitly mark a SoftwareComponent which is the Issuer and Audience of the user token
+- Add `AuthorizedFor` type to explicitly mark a SoftwareComponent which is part of the checked eligibility (permission group)
+- [**BC**] Use types above in `Authorize.authorizeAction` function
 
 ## 2.2.0 - 2021-02-16
 - Update dependencies